Moved DeepCopyPointer.h to DataFormats/Common
