+++ DataFormats/GeometryCommonDetAlgo/interface/GlobalErrorBaseExtended.h
@@ -1,7 +1,7 @@
 #ifndef GlobalErrorExtendedType_H
 #define GlobalErrorExtendedType_H
 
-#include "DataFormats/GeometryCommonDetAlgo/interface/DeepCopyPointer.h"
+#include "DataFormats/Common/interface/DeepCopyPointer.h"
 #include "DataFormats/Math/interface/AlgebraicROOTObjects.h"
 #include "DataFormats/GeometryVector/interface/GlobalPoint.h"
 //
