+++ DataFormats/GeometrySurface/interface/LocalErrorBaseExtended.h
@@ -1,7 +1,7 @@
 #ifndef LocalErrorType_H
 #define LocalErrorType_H
 
-#include "DataFormats/GeometryCommonDetAlgo/interface/DeepCopyPointer.h"
+#include "DataFormats/Common/interface/DeepCopyPointer.h"
 #include "DataFormats/GeometrySurface/interface/TrivialError.h"
 #include "DataFormats/Math/interface/AlgebraicROOTObjects.h"
 //
